// frontend/src/sanity/types.ts
export type Dish = {
  _id: string;
  _type: "dish";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  description?: string;
  price?: number;
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  restaurant?:
    | {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "restaurant";
      }
    | {
        _id: string;
        name?: string;
        logo?: SanityImage;
      };
  category?: "Burger" | "Sandwich" | "Fries" | "Pizza" | "Pasta" | "Salad" | "Sushi" | "Noodles" | "Rice" | "Coffee" | "Tea" | "Pastry" | "Drink" | "Dessert" | "Main Dish" | "Side Dish" | "Other";
};


export type CartItem = {
  _key: string;
  quantity: number;
  dish: {
    _id: string;
    name?: string;
    description?: string;
    price?: number;
    image?: SanityImage;
    restaurant?: {
      _id: string;
      name?: string;
      logo?: SanityImage;
    };
  };
};

export type CartOrder = {
  _id: string;
  items: CartItem[];
  totalPrice: number;
  status: string;
};

export type SanityImage = {
  _type: "image";
  asset: {
    _ref: string;
    _type: "reference";
  };
  crop?: {
    top: number;
    bottom: number;
    left: number;
    right: number;
  };
  hotspot?: {
    x: number;
    y: number;
    height: number;
    width: number;
  };
};

export type SanityReference = {
  _type: "reference";
  _ref: string;
  _key?: string;
};
///////////////////////////




export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;